+++ test/assets/module/util.ts
@@ -27,6 +27,7 @@ export const {
 	getMinMax,
 	getOption,
 	getPathBox,
+	getPointer,
 	getRandom,
 	getRange,
 	getRectSegList,